DataTagHDF5: Use new OutOfBounds exception
This is more appropriate then a plain runtime_error. We also
get rid of a bug where str + size_t was done, i.e. pointer
arithmetic was done, when the intention actually was to
concatenate the size_t to the string.
One part to silence the clang warnings about the aforementioned bug.