Fixes to DataArray class, new method in IDataArray and minor changes to testDataArray.
DataArray::createData now creates "data" field in "entity_group" (as it should). Furthermore "setExtent" now throws an exeption if the field is not set. Also a new method "hasData" has been implemented in the interface (and upwards) which is being used by the "setData" methods in "DataArray" to check if "createData" has already been called or not.