Made "id" be stored in "entity_group" but not in memory in HDF5
The base "Entity" now stores and gets "id" via "entity_group" and thus all getter-ctors and getters have been modified not to pass/require id.
The only exception is "Property" which would not support "setAttribute" on "property_group" since it uses "setData" on it. For Property a memory "id" has been implemented.