Made frontend class-type params be references. Split all multi-getter.
Multi-getter like "references" have now all been split into one without filter param and one with, to avoid default value. This was already done for a subset of multi-getters to allow passing them to validation and thus now is more consistent and allows to use references everywhere (which would not be possible for params with default value).