to_data_type: phenomenologically select DataType
We select the DataType, ie.g. float, based on the properties of the type (size, signedness, ...).
So even if two types are not same in the type system,
but have the same properties, e.g. long long unsigned and
long unsigned are both unsigned 64 bit integers, they are
now both correctly identified as uint64_t.
This should fix gh issue #486.