[DataFrame] more sugar for readColumn, writeColumn
Swap the arguments so both can be called without the offset
parameter. Separate the readColumn function to a simple one
without the count parameter (that can optionally resize
the vector it was given) and a extended one that wants
both count and resize parameters. Both functions have the
last parameter to specify an optional offset.
With that it is be possible to call:
  readColum(0, vals, [true])
  readColum(0, vals, count, <true|false>)
indicates the optional resize argument.