[dataAccess] dataSlice is now happy with less info ...
it is possible to provide only those dimensions which are actually sliced.
Information is applied first to last, e.g. if data is 3d and we give only
a one-dim position, it is applied to the first dim, the others are filled up
with defaults, i.e. offset or first element for start pos, last entry for
the end positions