[SectionHDF5] overload createProperty with shape arg ...
Change analogously to nixpy. A default shape of 8 entries is used for a
property that is created without a list of values. In previous
versions we let the automatic chunking define the shape which resulted
in a shape of 1024.

with '#' will be ignored, and an empty message aborts the commit.  # #
On branch property_size # Changes to be committed: # modified:
../backend/hdf5/SectionHDF5.cpp # modified:
../backend/hdf5/SectionHDF5.hpp # # Untracked files: # ../html/ #
../latex/ #